+++ signingscript/task.py
@@ -9,25 +9,27 @@ Attributes:
 """
 import aiohttp
 from frozendict import frozendict
-import json
 import logging
 import os
 import random
 import traceback
 
-import scriptworker.client
-from scriptworker.exceptions import ScriptWorkerException
-from scriptworker.utils import retry_request
+from datadog import statsd
+import platform
+
+from scriptworker.exceptions import ScriptWorkerException, TaskVerificationError
+from scriptworker.utils import retry_request, get_single_item_from_sequence
 
 from signingscript.sign import get_suitable_signing_servers, sign_gpg, \
     sign_jar, sign_macapp, sign_signcode, sign_widevine, sign_file
-from signingscript.exceptions import SigningServerError, TaskVerificationError
+from signingscript.exceptions import SigningServerError
 
 log = logging.getLogger(__name__)
 
 FORMAT_TO_SIGNING_FUNCTION = frozendict({
     "gpg": sign_gpg,
     "jar": sign_jar,
+    "focus-jar": sign_jar,
     "macapp": sign_macapp,
     "osslsigncode": sign_signcode,
     "sha2signcode": sign_signcode,
@@ -39,36 +41,58 @@ FORMAT_TO_SIGNING_FUNCTION = frozendict({
 })
 
 
+# task_cert_type {{{1
+def task_cert_type(context):
+    """Extract task certificate type.
+
+    Args:
+        context (Context): the signing context.
+
+    Raises:
+        TaskVerificationError: if the number of cert scopes is not 1.
+
+    Returns:
+        str: the cert type.
+
+    """
+    prefix = _get_cert_prefix(context)
+    scopes = context.task['scopes']
+    return get_single_item_from_sequence(
+        scopes,
+        condition=lambda scope: scope.startswith(prefix),
+        ErrorClass=TaskVerificationError,
+        no_item_error_message='No scope starting with "{}" found'.format(prefix),
+        too_many_item_error_message='More than one scope starting with "{}"'.format(prefix),
+    )
+
+
 # task_signing_formats {{{1
-def task_signing_formats(task):
+def task_signing_formats(context):
     """Get the list of signing formats from the task signing scopes.
 
     Args:
-        task (dict): the task definition.
+        context (Context): the signing context.
 
     Returns:
         list: the signing formats.
 
     """
-    return [s.split(":")[-1] for s in task["scopes"] if
-            s.startswith("project:releng:signing:format:")]
+    prefix = _get_format_prefix(context)
+    return [scope.split(':')[-1] for scope in context.task['scopes'] if scope.startswith(prefix)]
 
 
-# validate_task_schema {{{1
-def validate_task_schema(context):
-    """Validate the task json schema.
+def _get_cert_prefix(context):
+    return _get_scope_prefix(context, 'cert')
 
-    Args:
-        context (SigningContext): the signing context.
 
-    Raises:
-        ScriptWorkerTaxkException: on failed validation.
+def _get_format_prefix(context):
+    return _get_scope_prefix(context, 'format')
 
-    """
-    with open(context.config['schema_file']) as fh:
-        task_schema = json.load(fh)
-    log.debug(task_schema)
-    scriptworker.client.validate_json_schema(context.task, task_schema)
+
+def _get_scope_prefix(context, sub_namespace):
+    prefix = context.config['taskcluster_scope_prefix']
+    prefix = prefix if prefix.endswith(':') else '{}:'.format(prefix)
+    return '{}{}:'.format(prefix, sub_namespace)
 
 
 # get_token {{{1
@@ -76,7 +100,7 @@ async def get_token(context, output_file, cert_type, signing_formats):
     """Retrieve a token from the signingserver tied to my ip.
 
     Args:
-        context (SigningContext): the signing context
+        context (Context): the signing context
         output_file (str): the path to write the token to.
         cert_type (str): the cert type used to find an appropriate signing server
         signing_formats (list): the signing formats used to find an appropriate
@@ -119,7 +143,7 @@ async def sign(context, path, signing_formats):
     """Call the appropriate signing function per format, for a single file.
 
     Args:
-        context (SigningContext): the signing context
+        context (Context): the signing context
         path (str): the source file to sign
         signing_formats (list): the formats to sign with
 
@@ -135,7 +159,13 @@ async def sign(context, path, signing_formats):
             fmt, FORMAT_TO_SIGNING_FUNCTION['default']
         )
         log.info("sign(): Signing {} with {}...".format(output, fmt))
-        output = await signing_func(context, output, fmt)
+        metric_tags = [
+            'format:{}'.format(fmt),
+            'host:{}'.format(platform.node()),
+            'app:signingscript'
+        ]
+        with statsd.timed('signingfunc.time', tags=metric_tags):
+            output = await signing_func(context, output, fmt)
     # We want to return a list
     if not isinstance(output, (tuple, list)):
         output = [output]
@@ -174,7 +204,7 @@ def build_filelist_dict(context, all_signing_formats):
     a dictionary of `full_path` and signing `formats`.
 
     Args:
-        context (SigningContext): the signing context
+        context (Context): the signing context
         all_signing_formats (list): the superset of valid signing formats,
             based on the task scopes.  If the file signing formats are not
             a subset, throw an exception.